requirements.txt
pandas==3.0.1
pdfplumber==0.11.9
openpyxl==3.1.5
Pillow==12.2.0
